+++ TP3/src/raw.h
@@ -1,6 +1,6 @@
 #ifndef RAW_H
 #define RAW_H
 
-void generar_crudo(cimg_library::CImg<unsigned int>& orig);
+void generar_crudo(cimg_library::CImg<double>& orig);
 
 #endif
