+++ TP3/src/vecino.h
@@ -1,6 +1,6 @@
 #ifndef VECINO_H
 #define VECINO_H
 
-void vecino(cimg_library::CImg<unsigned int>& orig);
+void vecino(cimg_library::CImg<double>& orig);
 
 #endif
